requirements.txt
asgiref==3.11.0
Django
djangorestframework
psycopg2-binary==2.9.11
sqlparse==0.5.5
